requirements.txt
streamlit
pandas
google-generativeai
openai-whisper
pydub
SpeechRecognition
